refactor: extract encryption primitives and reorganize reader package
internal/encryption:
- Add package-level GoDoc to aad.go
- Add encrypt.go with EncryptGCM, SignGCM, EncryptCTR, EncodeModule
- Add ErrDecryptionFailed sentinel; DecryptGCM and VerifyGCMTag now
  return it on authentication failure instead of wrapping the internal
  crypto error
- Extract newGCMCipher helper, consolidating four cipher-creation sites
- Fix AES-CTR IV counter: use binary.BigEndian.PutUint32(iv[12:], 1)
  instead of setting only the last byte
- Add NIST SP 800-38D known-answer tests for DecryptGCM
- Add round-trip and large-payload tests covering all key sizes

reader:
- Drop redundant reader_ filename prefix from bloom, encryption,
  index, interop, options, and read_test files
- Dissolve reader_read.go: merge Read/ReadByNumber/ReadPartial/
  ReadPartialByNumber/read/fetchColumnData/unmarshalToResult/Reset/
  ReadStop into reader.go; move SkipRows into columnreader.go
- Remove ErrKeyRequired sentinel; use fmt.Errorf at each call site
  instead, consistent with the rest of the package
