+++ reader/bloom.go
