+++ reader/interop_test.go
