+++ reader/options.go
